feat: added notice banner component
show notice when conversation is in legacy mode, extracted some UI components from the ConversationHeader, updated featureReleaseTimestamp
