+++ ts/node/migration/sessionMigrations.ts
@@ -1225,7 +1225,8 @@ function updateToSessionSchemaVersion30(currentVersion: number, db: BetterSqlite
 
     // same value in ts/util/releaseFeature.ts but we cannot import since window doesn't exist yet.
     // TODO update to agreed value between platforms
-    const featureReleaseTimestamp = 1676851200000; // unix 13/02/2023
+    const featureReleaseTimestamp = 1677574800000; // unix 28/02/2023 09:00
+
     // support disppearing messages legacy mode until after the platform agreed timestamp
     if (Date.now() < featureReleaseTimestamp) {
       db.prepare(
